Rename triggers file
